you know what? let's try drawing it up with the tech on hand and THEN we'll componentize and do all that other crap
