+++ src/PokeAPI.js
@@ -15,7 +15,7 @@ function fetchJSON (endpoint) {
 /**
  * Get a list of all Pokemon species.
  */
-export async function getSpeciesList () {
+exports.getSpeciesList = async function () {
   const endpoint = `${baseURL}/pokemon-species/`
   // make the request
   return fetchJSON(endpoint)
@@ -27,7 +27,7 @@ export async function getSpeciesList () {
  * Get the information for a particular Pokemon species (e.g. Charizard, Dunsparce)
  * @param speciesName The name (or ID) of the species to get.
  */
-export async function getSpecies (speciesName) {
+exports.getSpecies = async function (speciesName) {
   const endpoint = `${baseURL}/pokemon-species/${speciesName}`
   return fetchJSON(endpoint)
 }
@@ -36,7 +36,7 @@ export async function getSpecies (speciesName) {
  * Get data for a particular Pokemon species variant (e.g. Sylveon, Mega Rayquaza)
  * @param pokemonName The name (or ID) of the Pokemon to get.
  */
-export async function getPokemon (pokemonName) {
+exports.getPokemon = async function (pokemonName) {
   const endpoint = `${baseURL}/pokemon/${pokemonName}`
   return fetchJSON(endpoint)
 }
