+++ src/components/Info.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+
+const Info = {}
+
+Info.Container = styled.div`
+  background-color: #1d1f21;
+  color: #d4d8db;
+  height: 100%;
+  width: 100%;
+`
+
+export default Info
