+++ src/index.js
@@ -1,12 +1,9 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import App from './App'
+import Dexplorer from './components/Main.js'
 import registerServiceWorker from './registerServiceWorker'
-import { BrowserRouter } from 'react-router-dom'
 
 ReactDOM.render(
-  <BrowserRouter>
-    <App />
-  </BrowserRouter>
+  <Dexplorer />
   , document.getElementById('root'))
 registerServiceWorker()
